refactor: 💡 export action types from actions.ts
