+++ lib/js/app/modules/app/actions.ts
@@ -219,3 +219,38 @@ export const copyApiResourceUrl = createAction(
     },
   })
 );
+
+export type AppActions =
+  | ReturnType<typeof appStart>
+  | ReturnType<typeof resizeScreen>
+  | ReturnType<typeof setScreenDimension>
+  | ReturnType<typeof shareQueryUrl>
+  | ReturnType<typeof editQuery>
+  | ReturnType<typeof queryEditorMounted>
+  | ReturnType<typeof notificationsMounted>
+  | ReturnType<typeof explorerMounted>
+  | ReturnType<typeof clearQuery>
+  | ReturnType<typeof createNewQuery>
+  | ReturnType<typeof updateQueryCreator>
+  | ReturnType<typeof setViewMode>
+  | ReturnType<typeof loadPersistedState>
+  | ReturnType<typeof showConfirmation>
+  | ReturnType<typeof hideConfirmation>
+  | ReturnType<typeof acceptConfirmation>
+  | ReturnType<typeof switchToQueriesList>
+  | ReturnType<typeof setVisualization>
+  | ReturnType<typeof resetVisualization>
+  | ReturnType<typeof showQuerySettingsModal>
+  | ReturnType<typeof hideQuerySettingsModal>
+  | ReturnType<typeof showEmbedModal>
+  | ReturnType<typeof hideEmbedModal>
+  | ReturnType<typeof copyEmbeddedCode>
+  | ReturnType<typeof downloadCodeSnippet>
+  | ReturnType<typeof selectFirstSavedQuery>
+  | ReturnType<typeof selectFirstSavedQuery>
+  | ReturnType<typeof showEmailExtractionModal>
+  | ReturnType<typeof hideEmailExtractionModal>
+  | ReturnType<typeof setQueryAutorun>
+  | ReturnType<typeof updateChartSettings>
+  | ReturnType<typeof updateVisualizationType>
+  | ReturnType<typeof copyApiResourceUrl>;
