+++ lib/js/app/modules/project/types.ts
@@ -1,7 +1,3 @@
-import { updateTagsPool } from './actions';
-
 export type ReducerState = {
   tagsPool: string[];
 };
-
-export type ProjectActions = ReturnType<typeof updateTagsPool>;
